Move autocomplete widgets to /autocomplete
