+++ ttkwidgets/autocomplete/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) The ttkwidgets authors 2017
+# Available under the license found in LICENSE
+from .autocomplete_entry import AutocompleteEntry
+from .autocompletecombobox import AutocompleteCombobox
