+++ ttkwidgets/autocomplete/autocomplete_entry.py
@@ -111,97 +111,3 @@ class AutocompleteEntry(ttk.Entry):
         """
         self.icursor(tk.END)
         self.selection_clear()
-
-
-class AutocompleteCombobox(ttk.Combobox):
-    """
-    Combobox widget that features autocomplete
-    """
-    def __init__(self, master=None, completevalues=None, **kwargs):
-        """
-        :param master: master widget
-        :param completevalues: autocompletion values list
-        :param kwargs: keyword arguments passed to Combobox initializer
-        """
-        ttk.Combobox.__init__(self, master, values=completevalues, **kwargs)
-        self._completion_list = completevalues
-        if isinstance(completevalues, list):
-            self.set_completion_list(completevalues)
-        self._hits = []
-        self._hit_index = 0
-        self.position = 0
-
-    def set_completion_list(self, completion_list):
-        """
-        Use our completion list as our drop down selection menu, arrows move through menu.
-        :param completion_list: list of values
-        :return: None
-        """
-        self._completion_list = sorted(completion_list, key=str.lower)  # Work with a sorted list
-        self._hits = []
-        self._hit_index = 0
-        self.position = 0
-        self.bind('<KeyRelease>', self.handle_keyrelease)
-        self['values'] = self._completion_list  # Setup our popup menu
-
-    def autocomplete(self, delta=0):
-        """
-        Autocomplete the Combobox, delta may be 0/1/-1 to cycle through possible hits
-        :param delta: int
-        :return: None
-        """
-        if delta:  # need to delete selection otherwise we would fix the current position
-            self.delete(self.position, tk.END)
-        else:  # set position to end so selection starts where textentry ended
-            self.position = len(self.get())
-        # collect hits
-        _hits = []
-        for element in self._completion_list:
-            if element.lower().startswith(self.get().lower()):  # Match case insensitively
-                _hits.append(element)
-        # if we have a new hit list, keep this in mind
-        if _hits != self._hits:
-            self._hit_index = 0
-            self._hits = _hits
-        # only allow cycling if we are in a known hit list
-        if _hits == self._hits and self._hits:
-            self._hit_index = (self._hit_index + delta) % len(self._hits)
-        # now finally perform the auto completion
-        if self._hits:
-            self.delete(0, tk.END)
-            self.insert(0, self._hits[self._hit_index])
-            self.select_range(self.position, tk.END)
-
-    def handle_keyrelease(self, event):
-        """
-        Event handler for the keyrelease event on this widget
-        :param event: Tkinter event
-        :return: None
-        """
-        if event.keysym == "BackSpace":
-            self.delete(self.index(tk.INSERT), tk.END)
-            self.position = self.index(tk.END)
-        if event.keysym == "Left":
-            if self.position < self.index(tk.END):  # delete the selection
-                self.delete(self.position, tk.END)
-            else:
-                self.position -= 1  # delete one character
-                self.delete(self.position, tk.END)
-        if event.keysym == "Right":
-            self.position = self.index(tk.END)  # go to end (no selection)
-        if event.keysym == "Return":
-            self.handle_return(None)
-            return
-        if len(event.keysym) == 1:
-            self.autocomplete()
-            # No need for up/down, we'll jump to the popup
-            # list at the position of the autocompletion
-
-    def handle_return(self, event):
-        """
-        Function to bind to the Enter/Return key so if Enter is pressed the selection is cleared
-        :param event: Tkinter event
-        :return: None
-        """
-        self.icursor(tk.END)
-        self.selection_clear()
